Changed filenames and locations to better reflect function
