+++ frontend/src/components/SingleAnswerSummary.tsx
@@ -8,7 +8,7 @@ import { MarkdownViewer } from '@equinor/fusion-components'
 import { Answer, Role } from '../api/models'
 import { organizationToString } from '../utils/EnumToString'
 import SeverityIndicator from './SeverityIndicator'
-import { useSharedFacilitatorAnswer } from './helpers'
+import { useSharedFacilitatorAnswer } from '../utils/helpers'
 
 interface SingleAnswerSummaryProps {
     answer: Answer
@@ -17,19 +17,16 @@ interface SingleAnswerSummaryProps {
 const SingleAnswerSummary = ({ answer }: SingleAnswerSummaryProps) => {
     const apiClients = useApiClients()
     const [username, setUsername] = useState<string>('')
-    const useShared = answer.answeredBy!.role === Role.Facilitator &&
-        useSharedFacilitatorAnswer(answer.progression)
+    const useShared = answer.answeredBy!.role === Role.Facilitator && useSharedFacilitatorAnswer(answer.progression)
     const organization = organizationToString(answer.answeredBy?.organization!)
 
     useEffect(() => {
         if (useShared) {
             setUsername('Facilitators')
-        }
-        else
-        {
-            apiClients.people.getPersonDetailsAsync(
-                answer.answeredBy!.azureUniqueId
-            ).then( details => { setUsername(details.data.name) })
+        } else {
+            apiClients.people.getPersonDetailsAsync(answer.answeredBy!.azureUniqueId).then(details => {
+                setUsername(details.data.name)
+            })
         }
     }, [])
 
@@ -43,7 +40,7 @@ const SingleAnswerSummary = ({ answer }: SingleAnswerSummaryProps) => {
                     <Box width={1}>
                         <Typography variant="h5">
                             {username}
-                            {useShared ? "" : ", ".concat(organization)}
+                            {useShared ? '' : ', '.concat(organization)}
                         </Typography>
                         <MarkdownViewer markdown={answer.text} />
                     </Box>
