+++ frontend/src/views/Evaluation/FollowUp/Views/Questionaire/FollowUpView.tsx
@@ -3,20 +3,20 @@ import React from 'react'
 import { Box } from '@material-ui/core'
 import { Button, Typography } from '@equinor/eds-core-react'
 
-import { Barrier, Evaluation, Question, Progression, Severity, Organization } from '../../../api/models'
-import AnswerSummarySidebar from '../../../components/AnswerSummarySidebar'
-import { barrierToString, progressionToString } from '../../../utils/EnumToString'
-import { useParticipant } from '../../../globals/contexts'
-import SeveritySummary from '../../../components/SeveritySummary'
-import QuestionProgressionFollowUpSidebar from './QuestionProgressionFollowUpSidebar'
-import { countSeverities } from '../../../utils/Severity'
-import QuestionsList from '../../../components/QuestionsList'
-import { useFilter } from '../../../utils/hooks'
-import OrganizationFilter from '../../../components/OrganizationFilter'
-import { getBarrierAnswers, onScroll } from '../../helpers'
-import { hasSeverity, hasOrganization, toggleFilter } from '../../../utils/QuestionAndAnswerUtils'
-import { disableAnswer, disableProgression } from '../../../utils/disableComponents'
-import { participantCanProgressEvaluation } from '../../../utils/RoleBasedAccess'
+import { Barrier, Evaluation, Question, Progression, Severity, Organization } from '../../../../../api/models'
+import { barrierToString, progressionToString } from '../../../../../utils/EnumToString'
+import { useParticipant } from '../../../../../globals/contexts'
+import { countSeverities } from '../../../../../utils/Severity'
+import { useFilter } from '../../../../../utils/hooks'
+import { getBarrierAnswers, onScroll } from '../../../../helpers'
+import { hasSeverity, hasOrganization, toggleFilter } from '../../../../../utils/QuestionAndAnswerUtils'
+import { disableAnswer, disableProgression } from '../../../../../utils/disableComponents'
+import { participantCanProgressEvaluation } from '../../../../../utils/RoleBasedAccess'
+import SeveritySummary from '../../../../../components/SeveritySummary'
+import QuestionProgressionFollowUpSidebar from '../../components/QuestionProgressionFollowUpSidebar'
+import QuestionsList from '../../../../../components/QuestionsList'
+import OrganizationFilter from '../../../../../components/OrganizationFilter'
+import AnswerSummarySidebar from '../../../../../components/AnswerSummarySidebar'
 
 interface FollowUpViewProps {
     evaluation: Evaluation
