+++ frontend/src/views/Evaluation/FollowUp/Views/WorkshopSummary/WorkshopSummaryView.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Box } from '@material-ui/core'
 
-import { Evaluation, Progression } from '../../../../api/models'
-import ListView from './ListView'
-import BowtieView from './BowtieView'
-import { AnswersWithBarrier } from '../../../../utils/Variables'
-import { assignAnswerToBarrierQuestions } from './helpers'
+import { Evaluation, Progression } from '../../../../../api/models'
+import { AnswersWithBarrier } from '../../../../../utils/Variables'
+import { assignAnswerToBarrierQuestions } from '../../util/helpers'
+import BowtieView from '../../components/BowtieView'
+import ListView from '../../components/ListView'
 
 interface Props {
     evaluation: Evaluation
