+++ frontend/src/views/Evaluation/FollowUp/components/BowtieView.tsx
