+++ frontend/src/views/Evaluation/FollowUp/util/helpers.ts
